product_scout_ai/requirements.txt
# Core dependencies
google-adk>=0.1.0
google-genai>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
click>=8.0.0
rich>=13.0.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.0.0

# Optional: API dependencies
fastapi>=0.100.0
uvicorn>=0.23.0

# UI dependencies
gradio>=4.0.0
plotly>=5.0.0
pandas>=2.0.0
